requirements.txt
azure-functions==1.21.3
pyodbc==5.2.0
